+++ frescobaldi_app/viewers/documents.py
@@ -61,13 +61,13 @@ def group(document):
 
 def load(filename):
     """Returns a Poppler.Document for the given filename, caching it (weakly).
-    
+
     Returns None if the document failed to load.
-    
+
     """
     mtime = os.path.getmtime(filename)
     key = (mtime, filename)
-    
+
     try:
         return _cache[key]
     except KeyError:
@@ -89,10 +89,10 @@ def filename(poppler_document):
 class Document(popplertools.Document):
     """Represents a (lazily) loaded PDF document."""
     updated = True
-    
+
     def load(self):
         return load(self.filename())
-        
+
     if popplerqt4 is None:
         def document(self):
             """Returns None because popplerqt4 is not available."""
@@ -101,19 +101,19 @@ class Document(popplertools.Document):
 
 class DocumentGroup(plugin.DocumentPlugin):
     """Represents a group of PDF documents, created by the text document it belongs to.
-    
+
     Multiple MusicView instances can use this group, they can store the positions
     of the Documents in the viewer themselves via a weak-key dictionary on the Document
     instances returned by documents(). On update() these Document instances will be reused.
-    
+
     The global documentUpdated(Document) signal will be emitted when the global
     app.jobFinished() signal causes a reload of documents in a group.
-    
+
     """
     def __init__(self, document):
         self._documents = None
         document.loaded.connect(self.update, -100)
-        
+
     def documents(self):
         """Returns the list of PDF Document objects created by our text document."""
         # If the list is asked for the very first time, update
@@ -121,19 +121,19 @@ class DocumentGroup(plugin.DocumentPlugin):
             self._documents = []
             self.update()
         return self._documents[:]
-    
+
     def update(self, newer=None):
         """Queries the resultfiles of this text document for PDF files and loads them.
-        
+
         Returns True if new documents were loaded.
         If newer is True, only PDF files newer than the source document are returned.
         If newer is False, all PDF files are returned.
         If newer is None (default), the setting from the configuration is used.
-        
+
         """
         if newer is None:
             newer = QSettings().value("musicview/newer_files_only", True, bool)
-        
+
         results = resultfiles.results(self.document())
         files = results.files(".pdf", newer)
         if files:
@@ -153,5 +153,3 @@ class DocumentGroup(plugin.DocumentPlugin):
                 documents.append(doc)
             self._documents = documents
             return True
-
-
