+++ frescobaldi_app/viewers/image.py
